Changing `lowdb` to `sqlite3`
i was using `lowdb` for reading data from json file.
but the data itself have relation, so i decided to
move the data to `sqlite3` by using `sequelize`
as ORM.

since the response object is little bit different,
i update `readme.md` so the response is match with
the documentation on `readme.md`.